+++ scripts/install/nginx/nginx-download.sh
@@ -23,8 +23,24 @@ fi
 # Start Main Script
 
 cd /usr/src
-rm -rf /usr/src/nginx-${NGINX_VER}
-wget https://nginx.org/download/nginx-${NGINX_VER}.tar.gz -O /usr/src/nginx-${NGINX_VER}.tar.gz --no-check-certificate && tar -xzvf /usr/src/nginx-${NGINX_VER}.tar.gz
-wget https://github.com/openresty/headers-more-nginx-module/archive/v${NGINX_HEADER_VER}.tar.gz -O /usr/src/v${NGINX_HEADER_VER}.tar.gz --no-check-certificate && tar -xzf /usr/src/v${NGINX_HEADER_VER}.tar.gz
-wget https://github.com/nginx-modules/ngx_cache_purge/archive/${NGINX_PURGE_VER}.tar.gz -O /usr/src/${NGINX_PURGE_VER}.tar.gz --no-check-certificate && tar -xzf /usr/src/${NGINX_PURGE_VER}.tar.gz
-wget https://github.com/openssl/openssl/releases/download/openssl-${OPENSSL_VER}/openssl-${OPENSSL_VER}.tar.gz -O /usr/src/openssl-${OPENSSL_VER}.tar.gz --no-check-certificate && tar -xzf /usr/src/openssl-${OPENSSL_VER}.tar.gz
+
+# Remove existing Nginx source directory if it exists
+if [ -d "/usr/src/nginx-${NGINX_VER}" ]; then
+  rm -rf /usr/src/nginx-${NGINX_VER}
+fi
+
+# Download and extract Nginx
+wget https://nginx.org/download/nginx-${NGINX_VER}.tar.gz -O /usr/src/nginx-${NGINX_VER}.tar.gz --no-check-certificate || { echo "Error: Failed to download Nginx."; exit 1; }
+tar -xzvf /usr/src/nginx-${NGINX_VER}.tar.gz || { echo "Error: Failed to extract Nginx."; exit 1; }
+
+# Download and extract Headers More module
+wget https://github.com/openresty/headers-more-nginx-module/archive/v${NGINX_HEADER_VER}.tar.gz -O /usr/src/v${NGINX_HEADER_VER}.tar.gz --no-check-certificate || { echo "Error: Failed to download Headers More module."; exit 1; }
+tar -xzf /usr/src/v${NGINX_HEADER_VER}.tar.gz || { echo "Error: Failed to extract Headers More module."; exit 1; }
+
+# Download and extract Cache Purge module
+wget https://github.com/nginx-modules/ngx_cache_purge/archive/${NGINX_PURGE_VER}.tar.gz -O /usr/src/${NGINX_PURGE_VER}.tar.gz --no-check-certificate || { echo "Error: Failed to download Cache Purge module."; exit 1; }
+tar -xzf /usr/src/${NGINX_PURGE_VER}.tar.gz || { echo "Error: Failed to extract Cache Purge module."; exit 1; }
+
+# Download and extract OpenSSL
+wget https://github.com/openssl/openssl/releases/download/openssl-${OPENSSL_VER}/openssl-${OPENSSL_VER}.tar.gz -O /usr/src/openssl-${OPENSSL_VER}.tar.gz --no-check-certificate || { echo "Error: Failed to download OpenSSL."; exit 1; }
+tar -xzf /usr/src/openssl-${OPENSSL_VER}.tar.gz || { echo "Error: Failed to extract OpenSSL."; exit 1; }
