+++ scripts/install/tools/system/admin-control-panel.sh
@@ -26,7 +26,12 @@ fi
 cd /usr/src
 
 # phpSysinfo
-rm -rf /var/www/admin/enginescript/phpsysinfo
+# Remove existing phpSysinfo directory if it exists
+if [ -d "/var/www/admin/enginescript/phpsysinfo" ]; then
+  rm -rf /var/www/admin/enginescript/phpsysinfo
+fi
+
+# Clone phpSysinfo
 git clone --depth 1 https://github.com/phpsysinfo/phpsysinfo.git /var/www/admin/enginescript/phpsysinfo
 cp -rf /usr/local/bin/enginescript/config/var/www/admin/phpsysinfo/phpsysinfo.ini /var/www/admin/enginescript/phpsysinfo/phpsysinfo.ini
 sed -i "s|SEDPHPVER|${PHP_VER}|g" /var/www/admin/enginescript/phpsysinfo/phpsysinfo.ini
