+++ scripts/install/tools/system/testssl-install.sh
@@ -25,8 +25,10 @@ fi
 # Return to /usr/src
 cd /usr/src
 
-# Remove Old Version
-rm -rf /usr/local/bin/testssl.sh
+# Remove existing Testssl.sh directory if it exists
+if [ -d "/usr/local/bin/testssl.sh" ]; then
+  rm -rf /usr/local/bin/testssl.sh
+fi
 
 # Install Testssl.sh
 git clone https://github.com/testssl/testssl.sh.git /usr/local/bin/testssl.sh
